refactor: _timerLoop を Clock に移動してリファクタ
