+++ lib/widgets/clock.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_pomodoro/models/dount_chart/chart_data.dart';
+import 'package:flutter_pomodoro/models/timer_loop.dart';
+import 'package:flutter_pomodoro/models/timer_mode.dart';
 import 'package:flutter_pomodoro/widgets/dounut_chart/donut_chart.dart';
 import 'package:flutter_pomodoro/widgets/timer_dial.dart';
 
@@ -11,6 +13,29 @@ class Clock extends StatefulWidget {
 }
 
 class _ClockState extends State<Clock> {
+  late TimerLoop _timerLoop;
+
+  @override
+  void initState() {
+    super.initState();
+
+    _timerLoop = TimerLoop(
+        timerMode: TimerMode.work,
+        onTick: () {
+          if (_timerLoop.remainingSeconds == 0) {
+            _timerLoop.switchAndRestart();
+          } else {
+            setState(() {
+              _timerLoop.decrementRemainingSeconds();
+            });
+          }
+        },
+        onEnd: () {
+          setState(() {});
+        });
+    _timerLoop.start();
+  }
+
   @override
   Widget build(BuildContext context) {
     return Stack(
@@ -24,8 +49,16 @@ class _ClockState extends State<Clock> {
           ],
           chartStrokeWidth: 10,
         ),
-        const TimerDial(),
+        TimerDial(
+          remainingSeconds: _timerLoop.remainingSeconds,
+        ),
       ],
     );
+  }
+
+  @override
+  void dispose() {
+    _timerLoop.dispose();
+    super.dispose();
   }
 }
